requirements.txt
# Kripto AL/SAT Tahmin Sistemi - Gerekli Kütüphaneler

# Veri işleme ve analiz
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Kripto veri çekme
python-binance>=1.0.19
ccxt>=4.1.0

# Teknik analiz göstergeleri
ta>=0.10.0

# Makine öğrenmesi ve tahminleme
scikit-learn>=1.3.0

# Zaman serisi analizi
statsmodels>=0.14.0

# Grafik ve görselleştirme
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Güvenlik ve şifreleme
cryptography>=41.0.0

# Loglama ve hata yönetimi
python-dotenv>=1.0.0

# Test ve geliştirme
pytest>=7.4.0
jupyter>=1.0.0 